Homeworks/Homework-6/requirements.txt
fastapi
uvicorn
pymongo
pydantic
requests
python-dotenv
